pyproject.toml
[project]
name = "vietnamese-vlm-dataset"
version = "0.1.0"
description = "Dataset generation pipeline for Vietnamese VLM fine-tuning"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "pillow>=10.0",
    "tqdm>=4.0",
    "python-dotenv>=1.0",
    "aiohttp>=3.9",
    "opencv-python>=4.8",
    "ffmpeg-python>=0.2",
    "albumentations>=1.3",
    "pycocotools>=2.0",
    "scipy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
